Move logic for loading RobotSettings to RobotSettings class
We need to use this logic in astprocd and astmetad now.
