+++ astoria/common/config/__init__.py
@@ -1,10 +1,11 @@
 """Configuration schemas for Astoria."""
 
 from .system import AstoriaConfig
-from .user import SSID_PREFIX, RobotSettings
+from .user import SSID_PREFIX, NoValidRobotSettingsException, RobotSettings
 
 __all__ = [
     "AstoriaConfig",
+    "NoValidRobotSettingsException",
     "RobotSettings",
     "SSID_PREFIX",
 ]
